refactor: move shared data
